+++ test/common/k8s/azureassignedidentity/azureassignedidentity.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"time"
 
-	aadpodid "github.com/Azure/aad-pod-identity/pkg/apis/aadpodidentity"
+	aadpodid "github.com/Azure/aad-pod-identity/pkg/internal"
 	"github.com/Azure/aad-pod-identity/test/common/util"
 	"github.com/pkg/errors"
 )
