+++ test/common/k8s/azureidentity/azureidentity.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 	"time"
 
-	aadpodid "github.com/Azure/aad-pod-identity/pkg/apis/aadpodidentity"
+	aadpodid "github.com/Azure/aad-pod-identity/pkg/internal"
 	"github.com/Azure/aad-pod-identity/test/common/util"
 	"github.com/pkg/errors"
 )
